Restructured ClientApp
