+++ GithubJobsEnterpriseProject/ClientApp/src/Images/empty-heart.png
