+++ GithubJobsEnterpriseProject/ClientApp/src/Images/full-heart.png
