+++ GithubJobsEnterpriseProject/ClientApp/src/Stylesheets/Job.css
@@ -4,7 +4,7 @@
 }
 
 .jobCard {
-    background: url("../cardBack.jpg");
+    background: url("../Images/cardBack.jpg");
     background-repeat: no-repeat;
     background-size: 350px 215px;
 }
