+++ GithubJobsEnterpriseProject/ClientApp/src/components/Home.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react'
-import Job from './Job';
+import Job from './JobComponents/Job';
 
 export default function Home(props) {
