+++ GithubJobsEnterpriseProject/ClientApp/src/components/JobComponents/JobDetailsDescriptionConverter.js
