+++ GithubJobsEnterpriseProject/ClientApp/src/components/JobComponents/SearchResults.js
